Add unit tests for the hop, state and verdict helpers
The sidecar had no test coverage, so a change to the hop-naming or the
hit/review thresholds could silently break the demo. These tests build a
tiny instructor pack in tmp_path and fake TypeSafe response objects, so
they run offline and never call Jev. Also wires ruff and pytest into CI,
and fixes the two lint findings ruff surfaced in the existing code.

+++ judge_lab.py
@@ -227,7 +227,7 @@ def _verdict(response: object) -> dict[str, object]:
         "Names the sink or the entry but not the connecting hop",
         "Names the entry, the identity hop, and the sink",
     )
-    idx = min(2, max(0, int(round(float(score.score)))))
+    idx = min(2, max(0, round(float(score.score))))
     uncertain = any(UNCERTAIN_LOW < p < UNCERTAIN_HIGH for p in (entry, hop, sink))
     return {
         "names_entry": entry,
